Align page layout with header, add nav/rating polish, and rework status menu
- Cap page content (Homepage panels, Library tabs/grid, BookChat) at the
  header's max-width so borders and dividers stop at the same edges instead
  of bleeding to the window edge.
- Redo Settings: replace personality emojis with icon wells, add a page
  subtitle.
- Add a shared theme.components.dropdown token set so floating menus (header
  account menu, BookChat status menu) share one visual language, documented
  as a standing rule in CLAUDE.md/.cursorrules.
- Add "Library" to the account dropdown so it's reachable from every page.
- Remove the book cover border and the redundant author/page byline in
  BookChat.
- Rework the "Currently Reading" status menu: fold page updates into the
  dropdown (no more clipped "Update page" widget), and fix a stacking-context
  bug where the menu's backdrop silently swallowed clicks on the menu itself.
- Support half-star ratings (tap left/right half of a star), backed by a
  Float rating column instead of Integer.

+++ src/services/db_models.py
@@ -53,7 +53,7 @@ class Book(Base):
     started_at = Column(String, nullable=True)
     finished_at = Column(String, nullable=True)
     status = Column(String, nullable=True)  # want_to_read | currently_reading | read | null
-    rating = Column(Integer, nullable=True)  # 1-5 | null
+    rating = Column(Float, nullable=True)  # 0.5-5 in half-star steps | null
 
     # Catch-all for any field the frontend sends that isn't modeled above yet
     # (the book shape has grown organically) — never silently drops data.
